--- src/pages/admin/index.tsx
@@ -1,3 +1,0 @@
-export const AdminPage = () => {
-  return <>admin</>;
-};
